python/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
discord.py==2.3.2
elevenlabs==1.6.1
executing==2.0.1
fonttools==4.47.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
ipython==8.22.2
jedi==0.19.1
joblib==1.3.2
kiwisolver==1.4.5
mariadb==1.1.10
matplotlib==3.8.2
matplotlib-inline==0.1.6
multidict==6.0.5
numpy==1.26.2
packaging==23.2
pandas==2.1.4
parso==0.8.3
Pillow==10.1.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.17.2
PyNaCl==1.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
threadpoolctl==3.2.0
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2023.3
urllib3==2.2.1
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4
